pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frappe-copilot-agent"
version = "0.1.0"
description = "AI agent service for Frappe/ERPNext"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=14.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "langchain>=0.3",
    "langgraph>=1.0",
    "langchain-openai>=0.3",
    "langchain-mcp-adapters>=0.1",
    "langgraph-checkpoint-postgres>=0.1",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "redis>=5.0",
    "alembic>=1.14",
    "pyjwt>=2.9",
    "structlog>=24.1",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48",
    "httpx>=0.27",
]

[project.optional-dependencies]
anthropic = ["langchain-anthropic>=0.3"]
google = ["langchain-google-genai>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-bdd>=7.0",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "pyright>=1.1",
    "fakeredis>=2.26",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "B", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires external services (PostgreSQL, Redis)",
]

[tool.hatch.build.targets.wheel]
packages = ["src/copilot_agent"]
